+++ test/models/sqlserver/object_default.rb
@@ -1,0 +1,3 @@
+class SSTestObjectDefault < ActiveRecord::Base
+  self.table_name = 'sst_defaultobjects'
+end