+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Improve case comparision performace per column. Fixes #414
 * DB rollback when reversable add_column has several options. Fixes #359
+* Better column definitions for default objects. Fixes #412
 
 
 ## v4.2.4